Major changes
Added email verification. It's also required to login. Once you make an account it automatically sends a verificationmail. Once your email is verificated you can login.  (This does mean you NEED an ACTIVE email account to create an account or test the registration/login)

Changed back button functionality.  You need to press the back button twice to close the app on the dashboard and login screen. If the back button is pressed on the register page, you'll be brought back to the login.

Added red required items to fields when left blank.

Changed cursor color on the input fields to make them vissible